requirements.txt
langchain
langchain-openai
langchain-groq
matplotlib
python-dotenv
pandas
streamlit